requirements.txt
# 核心文档处理库
pandas==2.1.0
openpyxl==3.1.2
python-pptx==0.6.21
python-docx==1.0.0

# PDF 处理
pdfplumber==0.10.2
PyPDF2==3.0.1

# 编码检测
chardet==5.2.0

# 数据可视化
matplotlib==3.8.0
seaborn==0.13.0

# 工具库
Pillow==10.1.0
python-dateutil==2.8.2

# 开发工具
pytest==7.4.3
black==23.11.0
